requirements.txt
attrs==20.3.0
docopt==0.6.2
flake8==3.9.0
iniconfig==1.1.1
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pycryptodome==3.10.1
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.3
toml==0.10.2
-e ./.